requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
oauthlib==3.1.0
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
SecretStorage==2.3.1
simplejson==3.16.0
sqlparse==0.4.1

